requirements/base.txt
django
psycopg2-binary
python-environ
django-cors-headers
django-jazzmin==3.0.1
drf-yasg==1.21.7
djangorestframework
django-filter==24.2
django-modeltranslation==0.19.14
django-redis==5.4.0
cryptography
django-recaptcha
nplusone
celery==5.5.1
redis
flower
dj-database-url
djangorestframework-simplejwt==5.2.2
django-phonenumber-field==8.1.0
django-phonenumbers==1.0.1
python-magic==0.4.27
Pillow==11.2.1
fcm-django==2.2.1
django-solo==2.4.0
django-celery-beat==2.8.1
jsonschema
pycountry==24.6.1
sentry-sdk
python-telegram-bot==13.15
django-auditlog==3.2.1
types-requests==2.32.4.20250913

django-versatileimagefield  
django-leaflet==0.32.0
